Use floats for duration; add float slider (why is this not in cloth config by default)

+++ src/main/java/moe/score/pishockzap/ZapController.java
@@ -58,7 +58,7 @@ public class ZapController {
 
                 // Waiting for shock to complete and then waiting for debounce time, so not a busy-wait per se
                 //noinspection BusyWait
-                Thread.sleep(config.getDebounceTime() * 1000L + (long) (shock.duration * 1000.0f));
+                Thread.sleep((long) (config.getDebounceTime() * 1000.0f) + (long) (shock.duration * 1000.0f));
             } catch (InterruptedException e) {
                 return;
             }
